chore: Move core functions into top level TTT folder
