+++ packages/tic-tac-toe/contracts/tests/tic-tac-toe.test.ts
@@ -1,3 +1,9 @@
+declare global {
+  interface Window {
+    ethers: any;
+  }
+}
+
 import {expectRevert, DeployedArtifact} from '@statechannels/devtools';
 import TicTacToeArtifact from '../../build/contracts/TicTacToe.json';
 import * as ethers from 'ethers';
@@ -14,7 +20,9 @@ import {
   randomExternalDestination
 } from '@statechannels/nitro-protocol';
 import {VariablePart} from '@statechannels/nitro-protocol';
-import {TTTData, PositionType, encodeTTTData} from './types/app-data';
+
+window.ethers = ethers; // Needed because core/app-data uses global ethers
+import {TTTData, PositionType, encodeTTTData} from '../../app/core/app-data';
 
 dotEnvExtended.load();
